chore: add small profile (timing) script
Run it with `make profile`, which runs scripts doing `hop.setup` with
different settings, and print the elapsed time (in nanoseconds).

This would be helpful to determine the cost of setup.

Currently mainly target on the `match_mappings` option, when all
mappings were loaded, it increases time about 0.3 ms on my machine.